+++ domain/crm/src/main/kotlin/com/few/crm/view/email/CrmEmailSendHistoryView.kt
@@ -1,7 +1,8 @@
-package com.few.crm.email.controller.view
+package com.few.crm.view.email
 
 import com.few.crm.email.domain.EmailSendHistory
 import com.few.crm.email.repository.EmailSendHistoryRepository
+import com.few.crm.view.CommonVerticalLayout
 import com.vaadin.flow.component.grid.Grid
 import com.vaadin.flow.component.orderedlayout.HorizontalLayout
 import com.vaadin.flow.component.textfield.TextArea
